refactor: extract global fee update logic into utility function
